+++ docs/CHANGELOG.md
@@ -12,3 +12,4 @@
 + Extend ST_CollectionExtract to filter with one or two dimensions
 + Add ST_Clip function to a [Multi]Polygon or [Multi]LineString geometry with another [Multi]Polygon or [Multi]LineString geometry
 + Add two utilities to return a list of numeric columns or the first one
++ Add ST_ForcePolygonCCW and ST_ForcePolygonCW functions